# apps/api/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/stores/{storeId}/products": {
      "post": {
        "operationId": "ProductsController_create",
        "parameters": [
          {
            "name": "storeId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateProductDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProductWithVariantsResponseDto"
                }
              }
            }
          }
        },
        "tags": [
          "Products"
        ]
      },
      "get": {
        "operationId": "ProductsController_findAll",
        "parameters": [
          {
            "name": "storeId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ProductDetailResponseDto"
                  }
                }
              }
            }
          }
        },
        "tags": [
          "Products"
        ]
      }
    },
    "/stores/{storeId}/products/{productId}": {
      "get": {
        "operationId": "ProductsController_findOne",
        "parameters": [
          {
            "name": "storeId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "productId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProductDetailResponseDto"
                }
              }
            }
          }
        },
        "tags": [
          "Products"
        ]
      },
      "patch": {
        "operationId": "ProductsController_update",
        "parameters": [
          {
            "name": "storeId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "productId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateProductDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProductResponseDto"
                }
              }
            }
          }
        },
        "tags": [
          "Products"
        ]
      },
      "delete": {
        "operationId": "ProductsController_softDelete",
        "parameters": [
          {
            "name": "storeId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "productId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProductResponseDto"
                }
              }
            }
          }
        },
        "tags": [
          "Products"
        ]
      }
    },
    "/stores/{storeId}/products/{productId}/publish": {
      "patch": {
        "operationId": "ProductsController_publish",
        "parameters": [
          {
            "name": "storeId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "productId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProductResponseDto"
                }
              }
            }
          }
        },
        "tags": [
          "Products"
        ]
      }
    },
    "/stores/{storeId}/products/{productId}/variants": {
      "post": {
        "operationId": "ProductsController_addVariant",
        "parameters": [
          {
            "name": "storeId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "productId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateVariantDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/VariantResponseDto"
                }
              }
            }
          }
        },
        "tags": [
          "Products"
        ]
      },
      "get": {
        "operationId": "ProductsController_listVariants",
        "parameters": [
          {
            "name": "storeId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "productId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/VariantResponseDto"
                  }
                }
              }
            }
          }
        },
        "tags": [
          "Products"
        ]
      }
    },
    "/stores/{storeId}/products/{productId}/variants/{variantId}": {
      "patch": {
        "operationId": "ProductsController_updateVariant",
        "parameters": [
          {
            "name": "storeId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "productId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "variantId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateVariantDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/VariantResponseDto"
                }
              }
            }
          }
        },
        "tags": [
          "Products"
        ]
      },
      "delete": {
        "operationId": "ProductsController_deleteVariant",
        "parameters": [
          {
            "name": "storeId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "productId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "variantId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/VariantResponseDto"
                }
              }
            }
          }
        },
        "tags": [
          "Products"
        ]
      }
    },
    "/stores/{storeId}/products/{productId}/images": {
      "post": {
        "operationId": "ProductsController_uploadImage",
        "parameters": [
          {
            "name": "storeId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "productId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "replace",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProductResponseDto"
                }
              }
            }
          }
        },
        "tags": [
          "Products"
        ]
      }
    },
    "/stores/{storeId}/products/{productId}/variants/{variantId}/images": {
      "post": {
        "operationId": "ProductsController_uploadVariantImage",
        "parameters": [
          {
            "name": "storeId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "productId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "variantId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/VariantResponseDto"
                }
              }
            }
          }
        },
        "tags": [
          "Products"
        ]
      }
    },
...
    "/stores/{storeId}/collections/{collectionId}/products/{productId}": {
      "delete": {
        "operationId": "CollectionsController_removeProduct",
        "parameters": [
          {
            "name": "storeId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "collectionId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "productId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CollectionProductResponseDto"
                }
              }
            }
          }
        },
        "tags": [
          "Collections"
        ]
      }
    },
...
  },
  "info": {
    "title": "Bias Market API",
    "description": "Bias Market — niche-first store builder API",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "schemas": {
...
      "CreateVariantDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "minLength": 1
          },
          "stock": {
            "type": "number"
          },
          "priceOverride": {
            "type": "number"
          },
          "imageOverride": {
            "type": "string"
          },
          "attributes": {
            "type": "object",
            "additionalProperties": {
              "type": "string"
            }
          }
        },
        "required": [
          "name"
        ]
      },
      "CreateProductDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "minLength": 2
          },
          "categoryIds": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "description": {
            "type": "string"
          },
          "price": {
            "type": "number"
          },
          "soldOut": {
            "type": "boolean"
          },
          "images": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "availableUntil": {
            "type": "string"
          },
          "currency": {
            "type": "string"
          },
          "stock": {
            "type": "number",
            "minimum": 0
          },
          "variants": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/CreateVariantDto"
            }
          }
        },
        "required": [
          "name",
          "price"
        ]
      },
      "VariantResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "productId": {
            "type": "string"
          },
          "storeId": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "stock": {
            "type": "number",
            "nullable": true
          },
          "reserved": {
            "type": "number"
          },
          "priceOverride": {
            "type": "string",
            "nullable": true
          },
          "imageOverride": {
            "type": "string",
            "nullable": true
          },
          "attributes": {
            "type": "object",
            "additionalProperties": {
              "type": "string"
            }
          }
        },
        "required": [
          "id",
          "productId",
          "storeId",
          "name",
          "stock",
          "reserved",
          "priceOverride",
          "imageOverride",
          "attributes"
        ]
      },
      "ProductWithVariantsResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "storeId": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "price": {
            "type": "string"
          },
          "currency": {
            "type": "string"
          },
          "images": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "availableUntil": {
            "type": "string",
            "nullable": true,
            "format": "date-time"
          },
          "status": {
            "enum": [
              "DRAFT",
              "PUBLISHED"
            ],
            "type": "string"
          },
          "soldOut": {
            "type": "boolean"
          },
          "deletedAt": {
            "type": "string",
            "nullable": true,
            "format": "date-time"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "variants": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/VariantResponseDto"
            }
          }
        },
        "required": [
          "id",
          "storeId",
          "name",
          "description",
          "price",
          "currency",
          "images",
          "availableUntil",
          "status",
          "soldOut",
          "deletedAt",
          "createdAt",
          "variants"
        ]
      },
      "CategoryInProductResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "name": {
            "type": "string"
          }
        },
        "required": [
          "id",
          "name"
        ]
      },
      "ProductCategoryResponseDto": {
        "type": "object",
        "properties": {
          "productId": {
            "type": "string"
          },
          "categoryId": {
            "type": "string"
          },
          "category": {
            "$ref": "#/components/schemas/CategoryInProductResponseDto"
          }
        },
        "required": [
          "productId",
          "categoryId",
          "category"
        ]
      },
      "ProductDetailResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "storeId": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "price": {
            "type": "string"
          },
          "currency": {
            "type": "string"
          },
          "images": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "availableUntil": {
            "type": "string",
            "nullable": true,
            "format": "date-time"
          },
          "status": {
            "enum": [
              "DRAFT",
              "PUBLISHED"
            ],
            "type": "string"
          },
          "soldOut": {
            "type": "boolean"
          },
          "deletedAt": {
            "type": "string",
            "nullable": true,
            "format": "date-time"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "variants": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/VariantResponseDto"
            }
          },
          "categories": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ProductCategoryResponseDto"
            }
          },
          "soldUnits": {
            "type": "number"
          },
          "availableStock": {
            "type": "number",
            "nullable": true
          }
        },
        "required": [
          "id",
          "storeId",
          "name",
          "description",
          "price",
          "currency",
          "images",
          "availableUntil",
          "status",
          "soldOut",
          "deletedAt",
          "createdAt",
          "variants",
          "categories",
          "soldUnits",
          "availableStock"
        ]
      },
      "UpdateProductDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "minLength": 2
          },
          "categoryIds": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "description": {
            "type": "string"
          },
          "price": {
            "type": "number"
          },
          "soldOut": {
            "type": "boolean"
          },
          "images": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "availableUntil": {
            "type": "string"
          },
          "currency": {
            "type": "string"
          }
        }
      },
      "ProductResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "storeId": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "price": {
            "type": "string"
          },
          "currency": {
            "type": "string"
          },
          "images": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "availableUntil": {
            "type": "string",
            "nullable": true,
            "format": "date-time"
          },
          "status": {
            "enum": [
              "DRAFT",
              "PUBLISHED"
            ],
            "type": "string"
          },
          "soldOut": {
            "type": "boolean"
          },
          "deletedAt": {
            "type": "string",
            "nullable": true,
            "format": "date-time"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          }
        },
        "required": [
          "id",
          "storeId",
          "name",
          "description",
          "price",
          "currency",
          "images",
          "availableUntil",
          "status",
          "soldOut",
          "deletedAt",
          "createdAt"
        ]
      },
      "UpdateVariantDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "minLength": 1
          },
          "stock": {
            "type": "number",
            "nullable": true
          },
          "priceOverride": {
            "type": "number",
            "nullable": true
          },
          "imageOverride": {
            "type": "string",
            "nullable": true
          },
          "attributes": {
            "type": "object",
            "additionalProperties": {
              "type": "string"
            }
          }
        }
      },
...
      "CollectionProductResponseDto": {
        "type": "object",
        "properties": {
          "collectionId": {
            "type": "string"
          },
          "productId": {
            "type": "string"
          },
          "position": {
            "type": "number"
          }
        },
        "required": [
          "collectionId",
          "productId",
          "position"
        ]
      },
...
    }
  }
}